requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
Django==3.0.3
djangorestframework==3.11.0
importlib-metadata==1.5.0
keyring==21.1.0
kombu==4.6.8
pytz==2019.3
pywin32-ctypes==0.2.0
sqlparse==0.3.0
vine==1.3.0
zipp==2.2.0
psycopg2
boto3
git+git://github.com/pymole/unisgf@master#egg=unisgf